refactor(build-logic): move and extend `libs` import to extension package
Refactored the `libs` import from the convention plugins to the `extensions` package for better organization. Key changes:
- Moved `libs` import to `com.xbot.convention.extensions.libs`.
- Added utility functions in `DependencyHandlerScope` for adding various types of dependencies.
- Extended `Project.libs` with additional documentation and structured handling.
