+++ build-logic/convention/src/main/kotlin/AndroidHiltConventionPlugin.kt
@@ -1,4 +1,4 @@
-import com.xbot.convention.libs
+import com.xbot.convention.extensions.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
