+++ build-logic/convention/src/main/kotlin/AndroidRoomConventionPlugin.kt
@@ -1,6 +1,6 @@
 import androidx.room.gradle.RoomExtension
 import com.google.devtools.ksp.gradle.KspExtension
-import com.xbot.convention.libs
+import com.xbot.convention.extensions.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
