+++ build-logic/convention/src/main/kotlin/com/xbot/convention/AndroidCompose.kt
@@ -3,6 +3,7 @@
 package com.xbot.convention
 
 import com.android.build.api.dsl.CommonExtension
+import com.xbot.convention.extensions.libs
 import org.gradle.api.Project
 import org.gradle.api.provider.Provider
 import org.gradle.kotlin.dsl.assign
