+++ build-logic/convention/src/main/kotlin/com/xbot/convention/KotlinAndroid.kt
@@ -1,6 +1,7 @@
 package com.xbot.convention
 
 import com.android.build.api.dsl.CommonExtension
+import com.xbot.convention.extensions.libs
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaPluginExtension
